requirements.txt
beautifulsoup4==4.11.1
dash==2.7.0
dash_bootstrap_components==1.2.1
email_validator==1.3.0
gunicorn==20.1.0
mysqlclient==2.1.1
pandas==1.4.3
protobuf==4.21.9
python_dateutil==2.8.2
PyYAML==6.0
requests==2.25.1
# google_api_python_client==2.65.0
# google_auth_oauthlib==0.7.1